+++ src/Json.Schema.Validation.Cli/Json.Schema.Validation.Cli.nuspec
@@ -1,0 +1,39 @@
+﻿<?xml version="1.0"?>
+<package>
+  <metadata>
+    <id>Microsoft.Json.Schema.Validation</id>
+    <version>$version$</version>
+    <title>Microsoft Json Schema Validation</title>
+    <authors>microsoft</authors>
+    <owners>microsoft,tse-securitytools</owners>
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <description>A utility that validates files according to a JSON schema.</description>
+    <releaseNotes>Version $version$ of the JSON validation utility (for SARIF v2.0.0)</releaseNotes>
+    <copyright>© Microsoft Corporation. All rights reserved.</copyright>
+    <licenseUrl>https://github.com/Microsoft/jschema/blob/master/LICENSE</licenseUrl>
+    <projectUrl>https://github.com/microsoft/jschema</projectUrl>
+    <iconUrl>https://go.microsoft.com/fwlink/?linkid=2009431</iconUrl>
+    <tags>Microsoft Json Schema .NET Validation Validator JSchema</tags>
+  </metadata>
+  <files>
+    <file src="bld\bin\$platform$_$configuration$\Json.Schema.Validation\net461\Microsoft.Json.Schema.Validation.dll"
+          target="lib\net461\Microsoft.Json.Schema.Validation.dll"
+          />
+
+    <file src="bld\bin\$platform$_$configuration$\Json.Schema.Validation\netstandard2.0\Microsoft.Json.Schema.Validation.dll"
+          target="lib\netstandard2.0\Microsoft.Json.Schema.Validation.dll"
+          />
+
+    <file src="bld\bin\$platform$_$configuration$\Json.Schema.Validation.Cli\net461\**"
+          target="tools\net461"
+          />
+
+    <file src="bld\bin\$platform$_$configuration$\Json.Schema.Validation.Cli\netcoreapp2.0\**"
+          target="tools\netcoreapp2.0"
+          />
+
+    <file src="src\ReleaseHistory.md" />
+    <file src="src\Json.Schema.Validation\**\*.cs" target="src\Json.Schema.Validation" />
+    <file src="src\Json.Schema.Validation.Cli\**\*.cs" target="src\Json.Schema.Validation.Cli" />
+  </files>
+</package>
