--- src/JsonSchemaValidator/JsonSchemaValidator.csproj
@@ -1,32 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFrameworks>netcoreapp2.0</TargetFrameworks>
-    <AssemblyNameOverride>JsonSchemaValidator</AssemblyNameOverride>
-    <RootNamespaceOverride>Microsoft.Json.Schema.Validation.CommandLine</RootNamespaceOverride>
-  </PropertyGroup>
-
-  <PropertyGroup Label="AssemblyAttributes">
-    <AssemblyTitle>JSON Schema Validation Tool</AssemblyTitle>
-    <Description>Command line tool to validate a JSON instance against a JSON Schema</Description>
-  </PropertyGroup>
-
-  <PropertyGroup Label="Package">
-    <IsTool>true</IsTool>
-  </PropertyGroup>
-
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), build.props))\build.props" />
-
-  <ItemGroup>
-    <PackageReference Include="CommandLineParser" Version="2.2.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
-    <PackageReference Include="Sarif.Sdk" Version="2.0.0-csd.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Json.Schema\Json.Schema.csproj" />
-    <ProjectReference Include="..\Json.Schema.Validation\Json.Schema.Validation.csproj" />
-  </ItemGroup>
-
-</Project>
